# dist/openapi.yaml
openapi: 3.0.0
info:
  title: Recipes api
  description: API for recipes project
  version: 1.0.1
servers:
  # local:
  - url: "http://127.0.0.1:3000"
    # remote:
    # - url: "http://132.72.65.211"
    description: SwaggerHub API Auto Mocking hey
...
paths:
...
  /users/RecipeDB:
    post:
      tags:
      - user
      summary: Add a new receipe to the user's database.
      description: Add a new receipe to the user's database.
      operationId: addRecipeToDB
      requestBody:
        description: Receipe added
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Receipe'
      responses:
        "201":
          description: receipe created successfully
        "400":
          description: "invalid input, object invalid"
    get:
      tags:
        - user
      summary: Get an internal receipe by ID
      description: Retrieve a specific internal receipe stored in the user's database.
      operationId: getRecipeFromDB
      parameters:
        - name: recipeId
          in: query
          required: true
          schema:
            type: integer
          description: The ID of the internal recipe to retrieve
      responses:
        "200":
          description: Recipe retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Receipe'
        "400":
          description: Invalid recipe ID
        "404":
          description: Recipe not found
  /users/like:
      post:
        tags:
          - user
        summary: Like a recipe
        description: Add a like to a recipe by the logged-in user.
        operationId: likeRecipe
        requestBody:
          description: Recipe to like
          required: true
          content:
            application/json:
              schema:
                type: object
                properties:
                  recipeId:
                    type: integer
                    example: 12345
                required:
                  - recipeId
        responses:
          "200":
            description: Like added successfully
          "400":
            description: Invalid input
          "401":
            description: Unauthorized
        security:
          - cookieAuth: []
      delete:
        tags:
          - user
        summary: Unlike a recipe
        description: Remove a like from a recipe by the logged-in user.
        operationId: unlikeRecipe
        requestBody:
          description: Recipe to unlike
          required: true
          content:
            application/json:
              schema:
                type: object
                properties:
                  recipeId:
                    type: integer
                    example: 12345
                required:
                  - recipeId
        responses:
          "200":
            description: Like deleted successfully
          "400":
            description: Invalid input
          "401":
            description: Unauthorized
        security:
          - cookieAuth: []
...
  /users/userLikedRecipe:
    get:
      tags:
        - user
      summary: Check if user liked a recipe
      description: Check if the logged-in user has liked a specific recipe.
      operationId: isUserLikedRecipe
      parameters:
        - name: recipeId
          in: query
          required: true
          schema:
            type: integer
          description: The ID of the recipe
      responses:
        "200":
          description: Whether the user liked the recipe
          content:
            application/json:
              schema:
                type: boolean
                example: true
        "400":
          description: Invalid recipe ID
        "401":
          description: Unauthorized
      security:
        - cookieAuth: []
...
  /users/addToViewRecipe:
    post:
      tags:
        - user
      summary: Record a user's recipe view
      description: Records the viewing of a recipe by a user and maintains only their 3 most recent viewed recipes.
      operationId: viewRecipe
      requestBody:
        description: Information about the viewed recipe
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                recipeId:
                  type: integer
                  example: 642583
                internalRecipe:
                  type: boolean
                  example: true
              required:
                - recipeId
                - internalRecipe
      responses:
        "200":
          description: Recipe view recorded.
        "400":
          description: Invalid input, object invalid
        "500":
          description: Server error while recording recipe view
  /users/getLastViewedRecipes:
    get:
      tags:
        - user
      summary: Get the last 3 viewed recipes for a user
      description: Retrieve the last 3 viewed recipes for a user.
      operationId: getLastViewedRecipes
      responses:
        "200":
          description: Last 3 viewed recipes retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Receipe'
        "401":
          description: Unauthorized
  /users/favorites:
    post:
      tags:
        - user
      summary: Add a recipe to favorites
      description: Save a recipe to the logged-in user's list of favorite recipes.
      operationId: markAsFavorite
      requestBody:
        description: Recipe to add to favorites
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                recipeId:
                  type: integer
                  example: 12345
                internalRecipe:
                  type: boolean
                  example: false
      responses:
        "200":
          description: The recipe was successfully saved as favorite
        "400":
          description: Invalid input
        "401":
          description: Unauthorized
    get:
      tags:
        - user
      summary: Get user's favorite recipes
      description: Retrieve the list of favorite recipes for the logged-in user.
      operationId: getFavoriteRecipes
      responses:
        "200":
          description: List of favorite recipes retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Receipe'
        "401":
          description: Unauthorized
    delete:
      tags:
        - user
      summary: Remove a recipe from favorites
      description: Remove a recipe from the logged-in user's list of favorite recipes.
      operationId: removeFavoriteRecipe
      requestBody:
        description: Recipe to remove from favorites
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                recipeId:
                  type: integer
                  example: 12345
              required:
                - recipeId
      responses:
        "200":
          description: Recipe removed from favorites
        "400":
          description: Missing recipeId
        "401":
          description: Unauthorized
  /users/history:
    post:
      tags:
        - user
      summary: Add a recipe to user's view history
      description: Save a recipe to the logged-in user's view history.
      operationId: addToHistoryRecipes
      requestBody:
        description: Recipe to add to history
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                recipeId:
                  type: integer
                  example: 456
                internalRecipe:
                  type: boolean
                  example: true
      responses:
        "200":
          description: The recipe was successfully added to history
        "400":
          description: Invalid input
        "401":
          description: Unauthorized
        "500":
          description: Server error or database failure
    get:
      tags:
        - user
      summary: Get user's view history
      description: Retrieve the list of recipes viewed by the logged-in user.
      operationId: getHistoryRecipes
      responses:
        "200":
          description: List of viewed recipes retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    recipeId:
                      type: integer
                      example: 456
                    internalRecipe:
                      type: boolean
                      example: true
        "401":
          description: Unauthorized
...
components:
  schemas:
    Receipe:
      required:
      - summary
      - glutenFree
      - image
      - extendedIngredients
      - analyzedInstructions
      - title
      - readyInMinutes
      - servings
      - vegan
      - sourceName
      - vegetarian
      type: object
      properties:
        id:
          type: integer
          example: 10
        title:
          type: string
          example: Spaghetti Bolognese
        image:
          type: string
          format: URL
          example: https://example.com/images/spaghetti.jpg
        readyInMinutes:
          type: integer
          example: 30
        servings:
          type: integer
          example: 4
        vegetarian:
          type: boolean
          example: false
        vegan:
          type: boolean
          example: false
        glutenFree:
          type: string
          example: false
        summary:
          type: string
          example: A classic Italian pasta dish with a rich meat sauce.
        extendedIngredients:
          type: array
          description: Detailed list of ingredients with metadata, quantities, and units
          items:
            type: object
            properties:
              id:
                type: integer
                example: 12061
              aisle:
                type: string
                example: "Nuts"
              image:
                type: string
                example: "almonds.jpg"
              consistency:
                type: string
                example: "SOLID"
              name:
                type: string
                example: "almond cookies"
              nameClean:
                type: string
                example: "almond cookies"
              original:
                type: string
                example: "150 grams Almond cookies, crumbed"
              originalName:
                type: string
                example: "Almond cookies, crumbed"
              amount:
                type: number
                example: 150.0
              unit:
                type: string
                example: "grams"
              meta:
                type: array
                items:
                  type: string
                example: []
              measures:
                type: object
                properties:
                  us:
                    type: object
                    properties:
                      amount:
                        type: number
                        example: 5.291
                      unitShort:
                        type: string
                        example: "oz"
                      unitLong:
                        type: string
                        example: "ounces"
                  metric:
                    type: object
                    properties:
                      amount:
                        type: number
                        example: 150.0
                      unitShort:
                        type: string
                        example: "g"
                      unitLong:
                        type: string
                        example: "grams"
        analyzedInstructions:
          type: array
          description: "Structured step-by-step preparation instructions"
          items:
            type: object
            properties:
              name:
                type: string
              steps:
                type: array
                items:
                  type: object
                  properties:
                    number:
                      type: integer
                    step:
                      type: string
                    ingredients:
                      type: array
                      items:
                        type: object
                        properties:
                          id:
                            type: integer
                          name:
                            type: string
                          localizedName:
                            type: string
                          image:
                            type: string
                    equipment:
                      type: array
                      items:
                        type: object
                        properties:
                          id:
                            type: integer
                          name:
                            type: string
                          localizedName:
                            type: string
                          image:
                            type: string
                          temperature:
                            type: object
                            properties:
                              number:
                                type: number
                              unit:
                                type: string
                    length:
                      type: object
                      properties:
                        number:
                          type: integer
                        unit:
                          type: string
          example:
            - name: ""
              steps:
                - number: 1
                  step: "Toss the cookie crumbs into the melted butter..."
                  ingredients:
                    - id: 10018192
                      name: "cookie crumbs"
                      localizedName: "cookie crumbs"
                      image: ""
                    - id: 1001
                      name: "butter"
                      localizedName: "butter"
                      image: "butter-sliced.jpg"
                  equipment:
                    - id: 404628
                      name: "hand mixer"
                      localizedName: "hand mixer"
                      image: "https://spoonacular.com/cdn/equipment_100x100/hand-mixer.png"
        sourceName:
          type: string
          description: The user who created or added the recipe to the family collection.
          example: user101
        popularity:
          type: integer
          description: The number of likes the recipe has
          example: 100
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session